# Query data source for glob patterns