--- a/bin/nanoc
+++ b/bin/nanoc
@@ -6,4 +6,10 @@ if File.file?('Gemfile') && !defined?(Bundler)
   warn 'A Gemfile was detected, but Bundler is not loaded. This is probably not what you want. To run Nanoc with Bundler, use `bundle exec nanoc`.'
 end
 
-Nanoc::CLI.run(ARGV)
+require 'ruby-prof'
+
+result = RubyProf.profile do
+  Nanoc::CLI.run(ARGV)
+end
+printer = RubyProf::GraphHtmlPrinter.new(result)
+File.open('graph.html', 'w') { |io| printer.print(io, {}) }