--- a/lib/nanoc/base/repos/site_loader.rb
+++ b/lib/nanoc/base/repos/site_loader.rb
@@ -15,6 +15,8 @@ module Nanoc::Int
     private
 
     class DocumentSource
+      attr_reader :data_source
+
       def initialize(data_source)
         @data_source = data_source
       end