--- a/spec/nanoc/base/entities/identifiable_collection_spec.rb
+++ b/spec/nanoc/base/entities/identifiable_collection_spec.rb
@@ -1,0 +1,7 @@
+describe Nanoc::Int::IdentifiableCollection do
+  let(:identifiable_collection) do
+    described_class.new(config)
+  end
+
+  let(:config) { {} }
+end