--- a/spec/nanoc/base/views/identifiable_collection_view_spec.rb
+++ b/spec/nanoc/base/views/identifiable_collection_view_spec.rb
@@ -143,7 +143,7 @@ shared_examples 'an identifiable collection' do
       let(:arg) { '/*.css' }
 
       it 'contains views' do
-        expect(subject.size).to eql(2)
+        expect(subject.to_a.size).to eql(2)
         about_css = subject.find { |iv| iv.identifier == '/about.css' }
         style_css = subject.find { |iv| iv.identifier == '/style.css' }
         expect(about_css.class).to equal(view_class)