**EXPERIMENTAL** and horribly work-in-progress.

When a glob is given to `IdentifiableCollection#[]`, query the data source and glob efficiently.

No visible speed improvement so far.