@ddfreyne 2015-11-14
```ruby
Benchmark.bmbm do |x|
  x.report('select') do
    10000.times { @items.select { |i| i.identifier =~ '/doc/*' } }
  end
  x.report('glob') do
    10000.times { Dir.glob('content/doc/*') }
  end
end
```

```
Rehearsal ------------------------------------------
select   3.390000   0.010000   3.400000 (  3.404135)
glob     0.220000   0.260000   0.480000 (  0.486837)
--------------------------------- total: 3.880000sec

             user     system      total        real
select   3.230000   0.010000   3.240000 (  3.238574)
glob     0.240000   0.280000   0.520000 (  0.526747)
```