@ddfreyne 2015-11-14
GitHub developer site doesn’t have a significant speed-up. Three test runs:

```
Without globbing:
  Site compiled in 14.29s.
  Site compiled in 14.16s.
  Site compiled in 14.58s.

With globbing:
  Site compiled in 14.19s.
  Site compiled in 14.02s.
  Site compiled in 14.00s.
```