@ddfreyne 2015-11-20
The preprocessor complicates this, because it can generate items/layouts just like data sources do. Perhaps it makes sense for the preprocessor to be backed by a data source. (But since the preprocessor can also delete items/layouts, this might become tricky too.)