@ddfreyne 2015-12-02
Will close this for now. Needs more work.